squash! README.md: Update poky commit to thud-20.0.0
weston-conf.bb was broken out of weston*.bb in commit
4001e1adb21b9785a19669968f07166f5b8772cd of poky. This caused meta-ivi to
install a weston.ini in two packages when building for qemux86* (the only
targets poky installs a weston.ini for so that the fbdev-backend is used).

Copy weston.ini.in from weston/ivi-shell/weston.ini.in so it can be used
in weston-conf.bbappend. Also update names of template variables to match
what has been done upstream and remove redundant replacements.

Should perhaps look into simplifying this. E.g. only have a weston.ini
that we just install. For now, only make it work like it did before the
change in poky so that we can update to a later commit on the thud branch.